# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mini Event Management API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
        "/events": {
            "get": {
                "summary": "List upcoming events",
                "description": "List upcoming events",
                "operationId": "3befe2e8582683a8dc70f7b706a5a97d",
                "parameters": [
                    {
                        "name": "tz",
                        "in": "query",
                        "description": "IANA timezone (e.g., Asia/Kolkata)",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "summary": "Create event",
                "description": "Create an event",
                "operationId": "1ed79d9ab6d41b87387cad6237213e5d",
                "parameters": [
                    {
                        "name": "tz",
                        "in": "query",
                        "description": "IANA timezone for interpreting input times",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "start_time",
                                    "end_time",
                                    "max_capacity"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "location": {
                                        "type": "string"
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "max_capacity": {
                                        "type": "integer",
                                        "minimum": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/events/{id}": {
            "get": {
                "summary": "Get event",
                "description": "Show a single event",
                "operationId": "abf4f4bbfe75221b5bab720a701f17b8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "tz",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "put": {
                "summary": "Update event",
                "description": "Update event",
                "operationId": "ca1f19a988fe9b1ee4af2f4ff3e724fe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "tz",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "summary": "Delete event",
                "description": "Delete event",
                "operationId": "54be98df9381a8b89013d29e170a31fe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/events/{id}/attendees": {
            "get": {
                "summary": "List attendees",
                "description": "List attendees",
                "operationId": "bc31b76dbc5c04942bcf38e7478da9cc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/events/{id}/register": {
            "post": {
                "summary": "Register attendee",
                "description": "Register attendee",
                "operationId": "a5d40dda446717b3353446c2b166c02f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "409": {
                        "description": "Conflict"
                    }
                }
            }
        }
    }
}